neo4j to rdf.ts
