+++ src/prec-0/from-cypher.d.ts
@@ -1,0 +1,9 @@
+import { IdentityTo, CypherNode, CypherEdge } from './PGDefinitions';
+import { Driver } from 'neo4j-driver'
+
+declare function extractFromNeo4jProtocole(
+  driver: Driver
+): Promise<{
+  nodes: IdentityTo<CypherNode>,
+  edges: IdentityTo<CypherEdge>
+}>;
